+++ django_athm/exceptions.py
@@ -6,7 +6,3 @@ class ATHM_Error(Exception):
 
 class ATHM_RefundError(ATHM_Error):
     pass
-
-
-class ATHM_ReportError(ATHM_Error):
-    pass