+++ django_athm/signals.py
@@ -6,4 +6,3 @@ athm_response_received = django.dispatch.Signal()
 # Status specific signals
 athm_completed_response = django.dispatch.Signal()
 athm_cancelled_response = django.dispatch.Signal()
-athm_expired_response = django.dispatch.Signal()